rpc: use atomic operations to manipulate rpc_transport refcount
Acquiring reference of rpc-transport instance is in the critical
path of handling every rpc call. The use of mutex to protect reference
count for  rpc-transport will therefore introduce a lot of contection,
in the case of lots of servicing threads.

Use atomic operations from glibc should eliminate those locking
contentions.

Signed-off-by: Zhang Huan <zhanghuan@open-fs.com>