cluster/stripe: fix issue of hanging frames
There is a lock issue in the stripe xlator. It took the wrong lock
for a frame so that the call counter might not be updated correctly,
making fop never returned. This issue is easy to trigger when gluster
client is configured with multiple event threads and request size is
larger than stripe size, this multiple sub-fop return. Fix it by
taking the correct lock.

Change-Id: I227ad607b3193cc23db78f0774aadb2862f701a5
Signed-off-by: Zhang Huan <zhanghuan@open-fs.com>