cluster/dht: fix crash when deleting directories
In DHT, after locks on sub-volumesare acquired, it would do rmdir in the
following sequence,
1. do rmdir on all subvolumes except the hased in a loop;
2. after 1 is done, do rmdir on the hashed subvolume

The problem is #1 is doing a check to skip sending rmdir to hashed
subvolume in the loop. If the last subvolume to check is actually the
hashed one, and #2 is quickly done before the loop ends, accessing
shared data, which is destroyed by #2, in #1 would cause a crash.

Fix this by moving access of shared data, that is the call frame and its
local data out of the loop in #1.

Change-Id: I8db7cf7cb262d74efcb58eb00f02ea37df4be4e2
Signed-off-by: Zhang Huan <zhanghuan@open-fs.com>